requirements.txt
pillow
pytesseract
google-generativeai
keyring
requests
pynput